권한 추가

+++ src/main/java/com/bit2016/mysite/controller/BoardController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import com.bit2016.mysite.service.BoardService;
 import com.bit2016.mysite.vo.Board;
+import com.bit2016.mysite.vo.Users;
 
 @Controller
 @RequestMapping("/board")
@@ -57,15 +58,27 @@ public class BoardController {
 	}
 
 	@RequestMapping("/delete")
-	public String delete(@RequestParam(value = "no", required = true, defaultValue = "0") Long no,
+	public String delete(@RequestParam(value = "no", required = true, defaultValue = "1") Long no,
 			@RequestParam(value = "p", required = true, defaultValue = "1") int p,
-			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd) {
-		boardService.delete(no);
-		return "redirect:/board?p=" + p + "&kwd=" + kwd;
+			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, HttpSession session) {
+		Users authUser = (Users) session.getAttribute("authUser");
+
+		Board board = boardService.view(no);
+		if (authUser != null) {
+			if (board.getUserNo() == authUser.getNo()) {
+				boardService.delete(no);
+			} else {
+				return "redirect:/user/loginform";
+			}
+		} else {
+			return "redirect:/user/loginform";
+		}
+
+		return "redirect:/board?p=" + p + "&kwd=" + 1;
 	}
 
 	@RequestMapping(value = "/modify", method = RequestMethod.GET)
-	public String modify(@RequestParam(value = "no", required = true, defaultValue = "0") Long no,
+	public String modify(@RequestParam(value = "no", required = true, defaultValue = "1") Long no,
 			@RequestParam(value = "p", required = true, defaultValue = "1") int p,
 			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, Model model) {
 		HashMap<String, Object> hm = boardService.view(no, p);
@@ -77,10 +90,22 @@ public class BoardController {
 	@RequestMapping(value = "/modify", method = RequestMethod.POST)
 	public String modify(@ModelAttribute Board vo,
 			@RequestParam(value = "p", required = true, defaultValue = "1") int p,
-			@RequestParam(value = "no", required = true, defaultValue = "1") int no,
-			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd) {
-		boardService.update(vo);
+			@RequestParam(value = "no", required = true, defaultValue = "1") Long no,
+			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, HttpSession session) {
+		Users authUser = (Users) session.getAttribute("authUser");
+		Board board = boardService.view(no);
+		if (authUser != null) {
+			if (board.getUserNo() == authUser.getNo()) {
+				boardService.update(vo);
+			} else {
+				return "redirect:/user/loginform";
+			}
+		} else {
+			return "redirect:/user/loginform";
+		}
+
 		return "redirect:/board/view?no=" + no + "&p=" + p + "&kwd=" + kwd;
+
 	}
 
 	@RequestMapping(value = "/reply", method = RequestMethod.GET)
